[FLINK-30396][test]replace the tab indent with space
