+++ flink-table/flink-table-planner/src/test/scala/org/apache/flink/table/planner/plan/stream/sql/UnnestTest.scala
@@ -17,10 +17,25 @@
  */
 package org.apache.flink.table.planner.plan.stream.sql
 
+import org.apache.flink.api.scala._
+import org.apache.flink.table.api._
 import org.apache.flink.table.planner.plan.common.UnnestTestBase
 import org.apache.flink.table.planner.utils.TableTestUtil
+
+import org.junit.Test
 
 class UnnestTest extends UnnestTestBase(true) {
 
   override def getTableTestUtil: TableTestUtil = streamTestUtil()
+
+  @Test
+  def testUnnestWithLookupJoinHint(): Unit = {
+    util.addTableSource[(Int, Array[Int])]("T2", 'a, 'b)
+    val sql = "SELECT /*+ LOOKUP('table'='D', 'retry-predicate'='lookup_miss', " +
+      "'retry-strategy'='fixed_delay', 'fixed-delay'='155 ms', 'max-attempts'='10', " +
+      "'async'='true', 'output-mode'='allow_unordered','capacity'='1000', 'time-out'='300 s') */ " +
+      "T2.a FROM T2 CROSS JOIN UNNEST(T2.b) AS D(c)"
+
+    verifyPlan(sql)
+  }
 }
